+++ integration-tests/pom.xml
@@ -115,14 +115,14 @@
 
 
         <plugins>
-            <plugin>
-                <groupId>org.eclipse.jetty</groupId>
-                <artifactId>jetty-maven-plugin</artifactId>
-                <version>${jetty.plugin.version}</version>
-                <configuration>
-                    <scanIntervalSeconds>2</scanIntervalSeconds>
-                </configuration>
-            </plugin>
+            <!--<plugin>-->
+                <!--<groupId>org.eclipse.jetty</groupId>-->
+                <!--<artifactId>jetty-maven-plugin</artifactId>-->
+                <!--<version>${jetty.plugin.version}</version>-->
+                <!--<configuration>-->
+                    <!--<scanIntervalSeconds>2</scanIntervalSeconds>-->
+                <!--</configuration>-->
+            <!--</plugin>-->
             <!--<plugin>-->
                 <!--<groupId>com.vaadin</groupId>-->
                 <!--<artifactId>vaadin-maven-plugin</artifactId>-->
@@ -208,46 +208,46 @@
                 </pluginRepository>
             </pluginRepositories>
         </profile>
-        <profile>
-            <id>start-jetty-for-it</id>
-            <activation>
-                <property>
-                    <name>skipTests</name>
-                    <value>!true</value>
-                </property>
-            </activation>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.eclipse.jetty</groupId>
-                        <artifactId>jetty-maven-plugin</artifactId>
-                        <executions>
-                            <!-- start and stop jetty (running our app) when running integration
-                                 tests -->
-                            <execution>
-                                <id>start-jetty</id>
-                                <phase>pre-integration-test</phase>
+        <!--<profile>-->
+            <!--<id>start-jetty-for-it</id>-->
+            <!--<activation>-->
+                <!--<property>-->
+                    <!--<name>skipTests</name>-->
+                    <!--<value>!true</value>-->
+                <!--</property>-->
+            <!--</activation>-->
+            <!--<build>-->
+                <!--<plugins>-->
+                    <!--<plugin>-->
+                        <!--<groupId>org.eclipse.jetty</groupId>-->
+                        <!--<artifactId>jetty-maven-plugin</artifactId>-->
+                        <!--<executions>-->
+                            <!--&lt;!&ndash; start and stop jetty (running our app) when running integration-->
+                                 <!--tests &ndash;&gt;-->
+                            <!--<execution>-->
+                                <!--<id>start-jetty</id>-->
+                                <!--<phase>pre-integration-test</phase>-->
 
-                                <goals>
-                                    <goal>start</goal>
-                                </goals>
-                                <configuration>
-                                    <!--<daemon>true</daemon>-->
-                                    <scanIntervalSeconds>0</scanIntervalSeconds>
-                                </configuration>
-                            </execution>
-                            <execution>
-                                <id>stop-jetty</id>
-                                <phase>post-integration-test</phase>
-                                <goals>
-                                    <goal>stop</goal>
-                                </goals>
-                            </execution>
-                        </executions>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
+                                <!--<goals>-->
+                                    <!--<goal>start</goal>-->
+                                <!--</goals>-->
+                                <!--<configuration>-->
+                                    <!--&lt;!&ndash;<daemon>true</daemon>&ndash;&gt;-->
+                                    <!--<scanIntervalSeconds>0</scanIntervalSeconds>-->
+                                <!--</configuration>-->
+                            <!--</execution>-->
+                            <!--<execution>-->
+                                <!--<id>stop-jetty</id>-->
+                                <!--<phase>post-integration-test</phase>-->
+                                <!--<goals>-->
+                                    <!--<goal>stop</goal>-->
+                                <!--</goals>-->
+                            <!--</execution>-->
+                        <!--</executions>-->
+                    <!--</plugin>-->
+                <!--</plugins>-->
+            <!--</build>-->
+        <!--</profile>-->
     </profiles>
 
 </project>
