+++ integration-tests/src/main/java/org/rapidpm/vaadin/addon/board/bootstrap/core/DDIVaadinServlet.java
@@ -21,8 +21,6 @@ public abstract class DDIVaadinServlet extends VaadinServlet {
 
     }
 
-    //add Metrics here
-
     @Override
     protected VaadinServletService createServletService(
         final DeploymentConfiguration deploymentConfiguration)
