+++ integration-tests/src/test/java/junit/com/vaadin/addon/board/v001/ExampleUITest.java
@@ -1,14 +1,10 @@
 package junit.com.vaadin.addon.board.v001;
 
-import javax.ws.rs.client.Client;
-import javax.ws.rs.client.ClientBuilder;
-
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.rapidpm.vaadin.addon.board.bootstrap.core.JumpstartUIComponentFactory;
-import junit.com.vaadin.addon.board.BaseTestbenchTest;
-import junit.com.vaadin.addon.board.MicroserviceBaseTest;
+
 import com.vaadin.board.Board;
 import com.vaadin.server.VaadinRequest;
 import com.vaadin.testbench.elements.ButtonElement;
@@ -17,45 +13,56 @@ import com.vaadin.ui.Button;
 import com.vaadin.ui.ComponentContainer;
 import com.vaadin.ui.Label;
 import com.vaadin.ui.VerticalLayout;
+import junit.com.vaadin.addon.board.core.BaseTestbenchTest;
 
 public class ExampleUITest extends BaseTestbenchTest {
 
+    public static class MyUIComponentFactory implements JumpstartUIComponentFactory {
 
-  @Before
-  public void setUp() throws Exception {
-    super.setUp();
-    openURL();
-  }
+        @Override
+        public ComponentContainer createComponentToSetAsContent(final VaadinRequest vaadinRequest) {
+            System.out.println("createComponentToSetAsContent.vaadinRequest = " + vaadinRequest);
+            VerticalLayout layout = new VerticalLayout();
+            Board board = new Board();
+            board.setSizeFull();
+            Button btn1 = new Button("Button 1");
+            Button btn2 = new Button("Button 2");
+            Button btn3 = new Button("Button 3");
+            Button btn4 = new Button("Button 4");
+            board.addRow(btn1, btn2, btn3, btn4);
 
-  @Test
-  public void test001() {
-    ButtonElement btn1 = $(ButtonElement.class).get(0);
-    btn1.click();
-    LabelElement label = $(LabelElement.class).get(0);
-    Assert.assertEquals("Example", label.getText());
+            Label label = new Label();
+            btn1.addClickListener(e -> label.setValue("Example"));
 
+            layout.addComponents(board, label);
+            return layout;
+        }
+    }
 
-    Client client = ClientBuilder.newClient();
-    final String generateBasicReqURL = baseRestURL(ComponentSwitcherRest.class);
-    System.out.println("generateBasicReqURL = " + generateBasicReqURL);
-    String val = client
-        .target(generateBasicReqURL)
-        .request()
-        .get(String.class);
-    System.out.println("val = " + val);
-    client.close();
+    @Before
+    public void setUp()
+        throws Exception {
+        super.setUp();
+        openURL();
+    }
 
-  }
+    @Test
+    public void test001() {
+        ButtonElement btn1 = $(ButtonElement.class).get(0);
+        btn1.click();
+        LabelElement label = $(LabelElement.class).get(0);
+        Assert.assertEquals("Example", label.getText());
+    }
 
+    @Test
+    public void test002() {
+        ButtonElement btn1 = $(ButtonElement.class).get(0);
+        btn1.click();
+        LabelElement label = $(LabelElement.class).get(0);
 
-  @Test
-  public void test002() {
-    ButtonElement btn1 = $(ButtonElement.class).get(0);
-    btn1.click();
-    LabelElement label = $(LabelElement.class).get(0);
+        Assert.assertEquals("Example", label.getText());
+    }
 
-    Assert.assertEquals("Example", label.getText());
-  }
 
 
 }
